EASY-1504 Upload files.
This PR only includes uploading files via a PUT. It does *not* include adding files to a directory
using PUT. I have implemented the servlet-part as well; this may conflict with #29 by @jo-pol.
However, we need this to work on Tuesday 17 April, so I suggest we integrate this now and improve
it after the demo.

**Important: ServletEnhancedLogging messes up the file upload, so I have commented it out**. After
the demo we need to see if we can fix it, or another approach is needed for the problem it is trying
to solve (standard logging for all servlet requests and responses).